requirements.txt
a2a-sdk[http-server]>=0.3.0
python-dotenv>=1.0.0
openai>=1.0.0
uvicorn
httpx>=0.27.0
pydantic>=2.0.0
biopython>=1.83
flask>=3.0.0
flask-cors>=4.0.0
